- Fixed issue #144 - Test data can now be added in docker as well.
