+++ lib/testdata.js
@@ -1,6 +1,6 @@
 const { getConfig } = require('./common');
 const { initDb } = require('./db');
-const { fixProductDates, fixDiscountDates, fixProductIds } = require('../test/helper');
+const { fixProductDates, fixDiscountDates, fixProductIds } = require('../lib/testhelper');
 const fs = require('fs');
 const path = require('path');
